requirements.txt
flask
flask-sqlalchemy
python-dotenv
requests
sqlalchemy